Harden artifact preview surface and split viewer components
